add initial current env integ test
